Coverage is good
Owing to a weakness in JaCoCo for code coverage in Kotlin inline functions,
this commit removes `inline` from those functions.

This is something that I would hope the _compiler_ handles automatically, as
the compilers for "C" languages can do automatically, in which case the
`inline` keyword is a hint rather than a command.