English denominations
This is confusing for a non-British reader:
- pence refers to a value
- pennies refers to more than one coin

Also note that string representation of fractional denominations (such
as halfpence and farthings) uses the _solidus_&mdash;not the
slash&mdash;to separate numerator from denominator.

This is a partial commit for English (British) denominations.
To add more denominations needs a local code strategy for representation
in standard "XsYd" format: both pence and shillings are bases for
denominating coinage quanties.