Drop quinzième as an Norman unit
This is from the Normal (and modern) French meaning "one part in
fifteen". It appeared after the Norman conquest, but was never
considered an English subdision of units of measure. This word is an
argument for sexagesimal measures that many English systems used (at
least in part; pioneered by the Babylonians).